+++ CHANGELOG.md
@@ -25,10 +25,12 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 * Update contribution calculation methods in `MediaSummary` with
   `aggregate_times` parameter to support granular time.
-
 * Add a `new_data` argument to `analyzer.optimal_freq()`.
 * Refactor args in `create_optimization_grid` to be consistent with
- `optimize(...)`.
+  `optimize(...)`.
+* Add support for optimization with forecasted data.
+* Deprecate `get_historical_spend` for `get_aggregated_spend` with `new_data`
+  support.
 
 ## [1.0.7] - 2025-03-19